animatedScaleComponent: Do not render children if not visible.
If the component is not visible, then due to animated style it's scale
is 0 which is causing a sudden splash of the children for a moment. A
solution for such sudden splash was introduced in 0baca8f. Now using
the same logic make it generic so that it will be applicable to all
other components which are using `AnimatedScaleComponent`.

`AnimatedScaleComponent` is also used by `UnreadNotice`: In the
`AnimationScaleComponent` we have been taking help of `display`
property (which acts same as it does in CSS), for same purpose. So
when visible is false, we were making display none just after
animation ends. Thus making it's size 0.

Now when visible becomes false, we will not be rendering anything
inside animation component and size thus will be 0. Thus no need to
display propery.

Basically this is a alternative of using display property. So next
`display` can be removed.

It's advantage over that is, it will result in smooth transition.
Before when visible becomes false, it scale down but height/size
remains same (resulting in white space) and as soon as animations
get's over it change display to none thus resulting in a sudden change.