unread: Correctly notice mentions in new messages.
We are calculating the unread mentions, in order to show them in the
UI though we currently don't show them (should do that soon.)

Due to not exposing this information, we haven't noticed an issue -
the logic to determine if the user was mentioned in an incoming
message was incorrect.  In fact, there is no `is_mentioned` key on the
`message` object; the data is included in `flags`.