android: Bump targetSdkVersion to 26 (!)
This is not thoroughly tested, and there'll be important follow-up
work for us to do.  It causes all systems running newer versions of
Android than our current target to treat our app with a number of
behavior changes introduced by the versions since then, up to the new
target -- which means 6.0, 7.0, 7.1, and 8.0, aka Marshmallow, Nougat,
and Oreo.  That's years of changes, with known and, surely, unknown
changes that affect us.

But we don't have much choice, because beginning on 2018-11-01 --
next week -- this is a requirement for uploading new releases to the
Play Store.  I'm going to send this today to alpha to help us shake
out problems.