ImagePicker: Add dependency for image-picker library change to follow.
This commit adds a dependency required by the new image picker library
react-native-image-crop-picker we will be switching to in a following commit.
This dependency is required by the library to enable its image cropping functionality
and even though we dont have that feature enabled currently the app will not build
without it.

See the Post-Install steps in
https://github.com/ivpusic/react-native-image-crop-picker/blob/07d321e3bc279b0ad218817245264cda6a7c77cb/README.md
for details.

Note: We are skipping adding
---------
vectorDrawables.useSupportLibrary = true
---------
as stated by the Readme  because the app seems to build and work fine without it
and its something required by the image cropping feature which we don't have enabled.
If we enable that feature we will want to add this line as well.

Addtionally, we want to take care to read the "Production build" instructions
while building the app for production.