fetch: Remove unused message fetch code in fetchEssentialInitialData.
There is some code in `fetchEssentialInitialData` left over from a
different behavior of the app.  We used to store the navigation state,
and on app restart restore the same stack of screens.  If they'd been
looking at a narrow (anywhere in the navigation stack), we would fetch
the messages for the narrow here.

Now we always start at the main navigation screen, not at whatever
screen the user was on.  (Even if the user is following a
notification, we enter the narrow via `doNarrow`, which takes care of
causing a fetch.)  So:
 * `getTopMostNarrow` here should always returns `undefined`, because
   there are no narrows on the navigation stack;
 * and this `getMessages` call, and its `messageFetchStart` and
   `messageFetchComplete`, should never happen.

We remove all this code and simplify the function greatly.