flow: Belatedly update flow-typed libdefs.
This brings everything in flow-typed up to date, using
`flow-typed update`.

Because we've been letting these get stale, this combines several
kinds of changes:
* Adding and removing libdefs to reflect changes we've made to
  our dependencies.
* Rebuilding autogenerated libdefs for dependencies we've bumped
  our version of.
* Where the flow-typed repo has had updates to its hand-written
  libdefs for our dependencies, taking those updates.

One of those updates in the last category has a side effect that we
need to start explicitly importing a type which we'd previously gotten
away with referring to implicitly in one file, so do that.