android notif: Inline `getNotificationBuilder` helper from library.
Customizing by subclassing was never a great design for this library.
This is an example: it makes it impossible to adapt to the new (in
API 26, last year) `Notification.Builder` constructor that takes two
arguments.

So, unwind that bit of design -- and take the opportunity to cut out
a bit of our dependence on this library in general.