android notif: Fix notifications not working in Android emulator.
Fixes #3113.

The native code inside `react-native-notifications` varied its
behavior when reading the `gcm_id` value from the metadata.  Because
of this, in the emulator it was throwing a `MismatchSenderId` error.

From an upstream issue thread, we take the solution of changing
`android:value="your_sender_id\0"` to `android:value="your_sender_id\"`.
This works on both an emulated and a physical device, despite the
warning in the official documentation:
   // IMPORTANT: Leave the trailing \0 intact!!!

While here, clean up a bit: extract this value to a string resource
and refer to it by name.