ios: Add a "bridging header"
This file is automatically created by Xcode when we add a Swift file
to the project, to allow us to communicate with our old Objective-C
classes from our Swift classes.  The changes to the project file were
also done by the same automatic action.

The header file was created as empty; the contents were added by hand.
It contains the RN bridge files that are needed.