webiew: Improve how images are displayed
Increases the maximum height of the images in the message list.
From 30vh to 40vh (viewport height). That is more consistent with
other mobile apps' image sizing and makes sure that a typical 4x3
photo will show as large as possible (touching the left and right
end of the list)