upload: Do not include filename in uploaded images
This is consistent with all mobile apps I tested (FB Messenger,
Slack, Hangouts).

This makes the uploaded image more clean. We should strongly consider
this to be the default behavior for the web app, with potentially
adding the metadata in the rendered html that can be parsed and
displayed if needed.