Add a placeholder when uploading an image
Use a placeholder resembing an image. It is sized similar to images
(100% wide and at most 40vw high) and has the green Zulip color with
the text 'Uploading...' inside.

A potential better future version is to pick its color by averaging
the uploaded image colors, or even better to show a thumbnail.