webview: Add CSS for read/unread messages
Add CSS styles for 'read' and 'unread' messages in the WebView.

Unread messages have a vertical bar on the very left (in 'brand color')
Read messages do not have any visible indicator.  Their style is
intended to animate the 'unread vertical bar' away from the screen.

The delay of one second is intended to give the user a chance to see
the indicator before it starts fading out.