Send read updates from outside to the WebView
At this point we are handling messages being read by scrolling the
WebView. This works well and is simple, contained in the WebView.

If a message becomes 'read' from outside, like being read on a
different device, or if other outside-the-Webview actions we do in
the app like 'Mark messages read', the WebView still didn't know
about these.

This implements a new WebView message type (event type!) that lets
the WebView know new messages are marked read. What to send is
determined by diffing the previous and the current 'read' flags and
if there are differences send the messages Ids to the WebView.

The WebView handler of this event cleverly creates a selector that
would catch only these nodes that need updating and then iterates
over them and sets their `data-read` attribute to `true`.