android build: Make release bundler work with new Android plugin.
Without this change, a release build succeeds... but crashes on
startup because it can't find `index.android.bundle`, the file that
should contain all our JS.  It can't find that file for the excellent
reason that there is no such file in assets/ inside the APK, as there
previously was in release builds.  Presumably this is due to some
combination of the upgrades to RN, to the Android Gradle plugin, and
to other build-time dependencies which we've done since 19.2.102.

Some poking around finds that the bundle is getting placed in the same
directory as it was in previous release builds:
  app/build/intermediates/assets/release/
but other assets which are still making it into the APK are going into
a different directory:
  app/build/intermediates/merged_assets/release/mergeReleaseAssets/out
So I guess that's the new place that assets should go.

Further poking around finds where the old directory name is coming
from: `react.gradle` hardcodes it.  Fixed upstream in commit
facebook/react-native@e90319e9f (facebook/react-native#21782)...
though by hardcoding a different scheme, so that still seems
vulnerable to breaking on a future plugin version. 🤷

Anyway, for the present: fortunately, we can override the bad old
hardcode with our own hardcode in our config, as a workaround.
Do that.