android build: Switch CI from the debug build to the release build.
This is the one that's much more likely to get broken without us
noticing, because we routinely run debug builds in development
but can go days or weeks between doing a release build for upload.
It's also I think more susceptible to getting broken in the first
place, because it has a few additional moving parts, notably the
JS bundle.

We could do both, but that sounds slow and unlikely to add a lot of
information.

Ideally we'd go a step further, and do a quick smoketest of actually
running the app -- confirm it starts up without crashing, and gets at
least to the initial "find your server" screen where our actual
application code is clearly running.

This wasn't possible until just recently, because we didn't have a
reasonable way to do a release build without also signing the APK and
uploading sourcemaps to Sentry, both of which require keys.  Thanks to
the work in #2883, now we do!