ui: Style disabled buttons according to Material Design guidelines
That makes the button gray and the text more readable while disabled

Styling includes both primary button style:
https://material.io/design/components/buttons.html#contained-button

and secondary button style:
https://material.io/design/components/buttons.html#outlined-button

The mid grey colors and transparency amounts are chosen in such a way
that when layed on a white background they produce the same colors
(or nearly the same, with priority to readable percentages, not math)
as the ones in the style guide.

The transparency makes this adaptable to night theme.