devAuthScreen: Fix performance issue in scrolling list.
It is observed that scroll performance degrades if there are nested
scrolling component. In devAuthScreen, there is flatList which has
it's own scroll handelling and it should not be nested with Screen
scolling component. So disbale scrolling in the Screen component of
devAuthScreen. This is also in continuation with
b0730a0636862f09d195c33be09d47776e449080.