message list: Stop re-using MessageList's own props type.
Note how `Props` is no longer exported, demonstrating that nothing
elsewhere refers to it.

This is awkwardly wordier as it stands, but that can be fixed with
some further refactoring... and at least we're no longer passing to
the child arbitary props we might add on the parent, or pretending the
parent has props actually only needed or present on the child.