Add and configure 'react-native-gesture-handler'
Reaict Native Gesture Handler provides native-driven gesture management APIs
for building best possible touch-based experiences in React Native.

While we might need to use it directly sometime in the future, the need to add
it now is that React Navigation 3 depends on it.

This PR adds 'react-native-gesture-handler' as dependency and configures it for
use in both iOS and Android. It follows the documentation on how to do it here:
https://reactnavigation.org/docs/en/getting-started.html
https://kmagiera.github.io/react-native-gesture-handler/docs/getting-started.html

First, `react-native link` was used, but then manually a lot of code changes that
were cosmetic and line reordering were undone. After that, some manual changes
were implemented (according to the documentation)