status bar: Make tests for color/style logic be true unit tests.
These tests as they were involved a ton of references to logic
elsewhere, namely getTitleBackgroundColor, and even mock Redux states.
That other logic has its own tests, so these are much clearer -- and
vastly shorter! -- if we separate that out.  In fact, when the
getTitleBackgroundColor calls are all replaced with their correct
return values, most of these test cases turn out to be identical!

At the same time, it becomes clear that `getStatusBarStyle`'s tests
are incomplete.  Filling those out will be a separate commit.

To help make the boundary between these areas of code a dependable
one, also factor out the special value 'transparent', which the two
`getStatusBar*` helpers specifically look for in the return value of
getTitleBackgroundColor, as an explicit named constant.