navbar: Remove dead code from `ModalSearchNavBar`.
In the past we were using different type of navBar, there was a search
icon in the right of navBar and on clicking this icon we were showing
search input in the navBar else screen name was shown. But now we are
not using this style.

Currently `ModalSearchNavBar` is only used in `Screen`, which passes
only some of these props; so `title`, `searchBar`, `clearSearchInput`
are always undefined.  That enables a lot of simplifications in this
code: `showSearchInput` is always true, the enable/disable handlers
are never called, and the state never changes.  Cut all the dead code.