Fix, add a test, improve `getRecipientsInGroupNarrow`
While adding Flow types, I found a bug in `getRecipientsInGroupNarrow`.
If a user is not found by the `allUsers.find` call, it is replaced with
an empty array `[]`. This was likely intended to be an empty object.

Funnily, doing a `user.email` returns an `undefined` regardless if
`user` is an object or an array. Neither will satisfy a strictly
typing object as a `User` type. This is a good argument in favor of
the `NULL_USER` constant. For that reason, I do not fully agree with
deprecating it.

Instead of adding a dummy/null object we can filter the user out.
So do that, and add a test to confirm this works.