types: Give a better type to Dispatch.
This requires us to fix another spot where we were returning the
`dispatch` return value where `void` is expected; and also to
straighten out something in the flow-typed annotations for
react-navigation which seems to be a mistake there, and causes
baffling enormous error messages when unaddressed.

This takes us from about 7.0% uncovered to about 6.5% uncovered
in `yarn test:flow-coverage` results.