apiTypes: Make unused User.profile_data too narrow instead of too wide.
We don't have a proper type annotation for this written down.  But we
don't currently use it at all, either, so that's basically fine.  Just
flip the type annotation so that if we do start using it we'll get an
error reminding us to go find the actual type, rather than silently
write code with type errors and not get them flagged.