tools/tsort-modules: Quick script to examine import graph and Flow status.
See the next commit for motivation.

The real engine doing the hard work here is Flow, parsing and
resolving the imports.  The `jq` command is a handy tool for
processing JSON data, which Flow is kind enough to provide; and
`tsort` is an old GNU utility to do a topological sort.