flow: Make one event-action type exact, as a quick demo.
In general we want to make almost all our object types exact, with
narrow exceptions where there's a specific reason for an inexact
object type.

One area where Flow makes it a bit funky to do this is where there's a
base type and some other types that extend it -- like our ServerEvent
and the various Event*Action types.  The idiom we've used for these is
intersection types, but those can't work when the base type is exact
(after all, the point of an exact type is that objects with more
properties *don't* qualify).  So we need to write them another way.

Demonstrate the new, exact-type-compatible idiom with an example.
I learned this idiom from its usage within React Native, e.g. in
StyleSheetTypes.js.  Flow's syntax here is kind of funny-looking,
but it's straightforward once you see it -- and it sure gives us
a more helpful type.