emoji: Consolidate tiny emoji-data functions, and fix a vague name.
This makes things a bit less scattered through lots of places to look
to understand the code.  The entire resulting file is 30 lines long,
which fits very easily on a screen.

Also, while we're going and editing the handful of imports of
emojiMap, fix those to refer explicitly to nameToEmojiMap.