emoji: Cut a spurious test.
This test asserts that we give different behavior to realm emoji
versus Unicode emoji -- looking for a match anywhere in the name for
realm emoji, but only at the start for Unicode emoji.

Looking anywhere might be a good behavior, but it should clearly be
uniform, so that the user doesn't have to know or care whether a given
emoji was built in or added by one of their colleagues.  Remove this
test, to decouple adding that behavior (or any variant of it) from
adding support for realm emoji.