message list: Fix reactions to show original emoji, not a revised one.
Sometimes we change what emoji a given name refers to.  When we do,
we should render old reactions with the emoji the author chose -- not
whatever emoji we might now offer under that name.  This is basically
the reason we store the `emoji_code` on a reaction in the first place.

We'd gotten that wrong; fix it.

Fixes #3169.