message render: Simplify out the pseudo-message object.
There may be a place for a "message, digested for rendering" object
somewhere, to better separate message rendering from the server API.
The one we have does not seem like it's doing very much of that,
though.  Simplify the code, and also open up space for another such
type in the future, by taking it out.