Include `prettier` test in `yarn test`
Include the `yarn run test:prettier` test when running tests locally
with the default `yarn test`. This will help reduce the commits with
incorrect formatting (for these who run the command)

The order of the command is changed to be consistent with the
`test:full` command that is run on the CI server with the only
difference being we are not running Coveralls locally.