npmrc: Set script-shell to bash.
It turns out `npm` (and Yarn, following it) has a way to configure
what shell to use for the "scripts", i.e. aliases, in package.json!
That allows us to escape the cold dead embrace of CMD.EXE for the
sake of contributors using Windows.

We could advance all the way to the shell technology of the '80s by
setting the shell to `sh`, the same value that's already the default
on Linux and presumably macOS... or we can go even further and treat
ourselves to the technology of the '90s and even the '00s, by
setting it to `bash`.  Let's do that. :-)

(Mac users will have Bash 3.2 from 2006 unless they've installed
something newer, so at least for now we'll stick to the mid-'00s.
Windows users will have installed a modern Bash as part of Git,
and Linux users will have a modern Bash in their base system.)

This reverts part of commit b6e468ff4 "package.json: Edit scripts to
be compatible with Windows cmd.exe", leaving the parts that were
semantically an improvement in any case.