webview: Make all font sizes scalable
If we specify sizes in 'em' they become relative to the parent
font size. `body` has a `font-size: 15px` and we already use `em`s
for lots of sizes but to make everything perfectly scalable
replace any `font-size: ???px` with the euqivalent `em` vlaue.

The font sizes are not calculated exactly `x px / 15 = y em`.
The original font sizes were arbitrary whole numbers relative to
the base font size, so we are keeping the same logic and precision,
instead of 0.6666em we use 0.7em because it still looks fine and
does not suggest more precision or intent behind the value.