tests: Remove old "e2e" version of Travis config.
This was basically moved from `.travis.yml` in commit f9e45a914
last year; it hasn't been touched since, and I don't believe
it's been used.  It's gotten stale in a variety of ways.

It'd be useful to have some form of end-to-end testing in CI,
but when we do if we want to draw from this it'll be in version
control; in the tree, it just causes confusion.