Do not reset `narrows`, `messages`, and `flags` on a dead queue
There are arguments for and against the reset of this data.
We need to improve the way we 'rehydrate' this information into the
app and there are simpler ways (discard data, reload) and better,
long-term ones - smarter queue.

This change brings the behavior consistent with what happens if
the app was killed and restarted. It is important that we treat
both cases the same way, as the often substitute each other -
the app may be resumed with a dead queue and may try reconnecting,
or it might have been killed producing a full restart.