howto/release: Rename from release-guide.
Since we moved this doc under a new "howto" directory a few months ago
in b1bbc5408, the "guide" part of the path has been rather redundant;
rename it away.

This causes me to notice there's a link in the README; take that out,
because this is a page that's mainly useful for maintainers.