version: Bump version to 21.0.104.
We've been leaving these constant in master, and only updating them
as part of the release process.  That's slightly messy, but hasn't
been a live problem... until now that we have zulipchat.com
informing CompatibilityChecker that, on Android, that ancient
version is too old.

So, update to reflect the latest version.