Add a `replaceExisting` prop to message fetch actions
To be able to differentiate the intent of the message fetch actions
we add a new prop named `replaceExisting`.

Any fetch action needs to pass the `replaceExisting` down to the
dispatched `messageFetchComplete` action which is used to store
the messages.

Initially, I wanted to make this an optional function argument
for the action crators so it is simpler to integrate with existing
code. So I made it mandatory and also two other arguments `anchor`
and `useFirstUnread`.