input: Make props type exact, and resolve mystery.
As prophesied, Flow was indeed telling us about a real type error!
Discovered the cause -- this bogus `defaultProps` entry of
`restProps`, which also explains where Flow was getting that
identifier from for the error messages -- while making the cleanups
in the preceding few commits, which also involved `defaultProps`.

And now we happily get to make this type exact.  On attempting to
pass to Input a bogus prop that it doesn't actually accept, Flow now
successfully gives a type error.