PasswordInput: Pass through all props to Input again, but explicitly.
In 4dd4752d8 "login: Stop splatting props in PasswordInput." we
replaced a `{...this.props}` in the `<Input ... />` with explicitly
mentioning the only two props that the type annotation allowed for
this component: `style`, which the next commit e3e3821c8 then cut out
entirely, and `placeholder`.

In fact, though, at the one call site of this component, we were
passing several other props as well!  And all were intended to get
passed through to `Input`; when they weren't, that broke the
password-login screen entirely.  Oops.

Fix this two ways:

 * Pass all props through to `Input`.  Like `MultilineInput`, this
   is a wrapper around that component that exposes most of the
   latter's customization options by forwarding props.

 * Say that in the types and the jsdoc, rather than asserting the
   component takes only one or two of the props it actually takes.