i18n: Write down real type of LocalizableText; fix errors.
See preceding commit for how the `any` here gagged Flow from pointing
out to me a bad regression in a refactor.  (Fortunately so bad it'd
never have made it to users, but still a waste of time).

Even before today, the `any` was masking a type error in ModalNavBar:
the `title` prop wasn't really optional at all, and if a caller had
tried to leave it out, the app would have crashed.  So fix that.

And TranslatedText needs an adjustment similar to the one just made to
Input, to make explicit its intention to handle a heterogeneous type.