SearchInput: Drop never-used property `textInput`.
As far as I can tell, this property of this component has never been
used in our codebase's history, since 64a6ebb36 when it was added.
At some point it stopped getting actually written to, either, and the
parent commit removed the code in this file that appeared to do so.
Remove the property declaration too.