autocomplete: Fix popup sometimes not responding to touch.
Fixes #3209.

In 824c908cc we introduced a new behavior for `ComposeBox` to not
unmount when one can't write to a narrow.  I would not call it a
'hack' as this approach should have worked fine (and would on web).
But seemingly it is throwing React Native off, and showing the
Autocomplete visually z-above the WebView but z-below in the stack of
'touchable' targets.

Luckily, a better, less hacky approach is possible.  Just use a
familiar style from the web: `display: 'none'`.