api: Rename `focusPing` -> `reportPresence`.
I don't quite know what a "focus ping" would be, but it's not a term
in the Zulip protocol.

The system this request is part of is called "presence".  Just saying
`presence` feels too short and general; this particular request
reports the user's presence on this client (and serves a double role
by getting back updates on other users' presence), so name it by that.