reactions: Replace an intersection type that extra-confuses Flow.
In general Flow seems to have a hard time reasoning effectively about
intersection types.  This is an especially bad case; when I make a
sweep through our action types to make most of them exact, this one
line is the difference between "No errors!" and *ahem* "1 error"
that (command below edited slightly)...

  $ yarn flow | wc
     1048    8080   51937

takes over 1000 lines and 50kB to print.  "Found 1 error", it says.

In that enormous spew, the spot where it first got interesting and
relevant-looking was here, about line 228:

 • Or property id is missing in object type [5] but exists in
   object type [8] in type argument A.
 ...
     [3][8] 350│ export type EventReactionCommon = ServerEvent &
     [5][6] 351│   Reaction & {
            352│     message_id: number,
            353│   };

(In the original, the "8" and the two "&" are colored yellow,
contrasting with the blue of the rest -- I think the point is that
"[8]" refers to the whole "A & B & C" type-expression while "[3]",
"[5]", "[6]" are the A, B, C respectively.)

And indeed, we don't really want to say that one of these events
"is a" `Reaction` object -- so an intersection isn't appropriate here.
What we really want is just a shorthand for saying they carry the
same list of properties.  That's what a spread is for.