android notif: Handle all notif actions with our own intent service.
This lets us cut out the wix ProxyService entirely.

The new logic to construct the intent, seen in getNotificationBuilder,
is straight from upstream docs:
  https://developer.android.com/training/notify-user/build-notification#Actions

We use the standard Android action value ACTION_VIEW.  We also follow
standard Android idioms in using the "data" field, in URI format, to
identify the target / the thing to be viewed -- so that the intent's
"action", its "data", and its target component/class, together
describe what it is you want to do.

The data field isn't currently consumed by the app, though -- it just
relies on the "extra", which we continue to set in the same way.

On the service side, the new ACTION_VIEW block basically just inlines
for now what the wix code does: ProxyService#onHandleIntent, then two
method calls found there.