nav: Take off createSelector from getInitialNavState.
We run this once at rehydrate time; there's no benefit to caching,
so cut the complexity.

Also it isn't even really a selector, as its input is really the
state-like object provided at rehydration, not the Redux state...
we'll get to that next.

As a bonus, this function now gets type-checked well enough for Flow
to notice that it was passing a maybe type where it wasn't supposed to
go!  Fix that; the function actually is fine with null and undefined,
just needs to say so.