WIP webview: Make padding and margin not scalable with font size
[This commit has the rump of changes I think we don't want,
 here for temporary rebase convenience.  The changes described here
 were merged in previous commits, mainly ebaffb456.]

Change the units used for padding and margins from 'rem' to 'px'.
This makes these UI elements not scale with font size change.

While this would be undesirable from a purely designer's point of view,
and not something we would do for a web app, that is a best practice
that the tested apps follow on mobile. This leaves more space for
the text when enlarged, considering the limited space.

The values are not always mathematically the same, but the intention
is always preserved:
 * when possible: 1rem => 16px, 0.5rem => 8px, 0.25rem => 4px
 * 0.2rem => 4px not because of faulty math but for consistency