android notif: Implement group summary messaging styled notifications
Enhances the current inbox styled notifications into grouped messaging
style notifications.

This adds the ability to expand each notification to view all the
previous messages instead of the only last message of each narrow
which was being shown in the previous commits.

This is the basic functionality of the messaging styled notifications
more polishing to come in the subsequent commits.

Makes the action replies easier to implement which is a task of
future.