notif: Factor out "unregister" logic from two components.
This is better handled with our other notifications code than in
component code... and much better than in *two* components' code,
doing the exact same thing.

We'll move most of this logic to the actual `notifications.js` in the
next commit, like we did for the `initNotifications` action in a
recent previous commit; committing this separately to make super clear
the aspect of this change that is just moving this code verbatim.

The actual logic this is carrying out is pretty dubious, but we'll be
in a better position to change and fix it after moving it where it
belongs.