Screen: Adjust `children` prop's type to match ScrollView.
Some other changes I was experimenting with cause Flow to start
complaining at call sites of `Screen` that it can't figure out what
type to treat the children as.

What we do with `children` is pass it straight on to the `ScrollView`,
so I consulted that component's code... and it declares this prop as
having type `React.Node`.  (For which we have an alias `React$Node`
which simplifies importing it.)  So, match that.