account: Split `getActiveAccount`, with a version that asserts.
In the bulk of the places where we use this (mostly via `getAuth` and
a few other thin wrappers around it), we're rendering UI that is all
about messages, users, etc. that the user has access to via a
particular account -- so the code pervasively assumes, and should
assume, that there is in fact an active account.

So, provide a way to make that explicit.  No empty-string fallback
which might not crash but would look badly broken; no `undefined`
fallback which would propagate through the types and force a bunch of
never-exercised fallback code elsewhere, much of which again would
produce badly broken results if it ever did somehow fire; just
acknowledge the assumption up front, and then have an honest, valid
Account object.

This function itself has only a handful of call sites.  The one site
outside this file is one that wants the "assert" version, like most of
our code.  The others have a number of call sites of their own, so we
keep their semantics unchanged for now, and we'll handle them in
separate commits.