notif: Listen globally for device token, and send to all servers.
There's only one token, and it's independent of the Zulip server;
so we only need one listener to learn it, and store it in `session`
for everything else to use.  Do that in NotificationListener.

Then as soon as we learn it, or learn it's changed, there's no
reason to wait to tell every server that needs to know.  Make a new
action creator `sendAllPushToken`, and call that.

Once we're doing that, the one part of this we still may need to do
at login time (in getToken*) is tell the server a value that we've
already learned for the token.  So, call `maybeSendPushToken`
specifically.

This fixes #2128: we no longer accumulate ad-hoc listeners as the user
switches between accounts.

---

In comparing the old and new code, one helpful observation is that the
wix library's `setRegistrationTokenUpdateListener` just adds an event
listener on `remoteNotificationsRegistered`, i.e. the same event name
as the iOS side already refers to:

  $ rg -NA2 " setRegistra" node_modules/
  node_modules/react-native-notifications/index.android.js
    static setRegistrationTokenUpdateListener(listener) {
      registrationTokenUpdateListener = DeviceEventEmitter.addListener("remoteNotificationsRegistered", listener);
    }