android notif: Stop overriding our sound with DEFAULT_SOUND.
This has no visible effect right now, because `getNotificationSoundUri`
chooses the default notification sound anyway; but it makes that code
actually control the sound setting, untangling some confusion.  (Though
only on API <26; on API >=26, it's controlled entirely by the channel.)

We keep the other flags: DEFAULT_VIBRATE, with a TODO to evaluate the
custom vibration pattern we ineffectively attempt to set, and
DEFAULT_LIGHTS because we don't have any alternative to prefer.

Thanks to @jainkuniya for part of the investigation leading to this fix.