session: Cut extra indirections in action creators.
This is a pure refactor: the only effect our reducers give to these
actions is to update these corresponding bits of state.  So the only
possible value in having these checks is as an optimization.

It's possible there are cases where a pattern like this would be a
helpful optimization; but these action creators are all invoked very
infrequently, and even less often will they be invoked when the value
hasn't changed.  So just drop this complexity.