redux: Inline some more tiny handlers into reducers.
With this and the parent commit, we're down from about
114 out-of-line handler functions to just 41 -- those that have
the most logic in them and don't nicely fit inline in the reducer.

Most of those 41 aren't fundamentally that complicated, either --
they become complicated only because of the particular data
structures we happen to be using for those bits of the state, or
because of the lack of certain common abstractions.  Now that the
boilerplate for an action handler that really is very simple has
been greatly reduced, there's an opportunity to refactor many of
these to take advantage of that.