webview: Add 'react-native-webview' dependency.
The `WebView` component is being removed out of RN Core
as part of the "Lean Core" project:
  https://github.com/facebook/react-native/issues/23313

Its development continues here:
  https://github.com/react-native-community/react-native-webview

Just add the dependency to `package.json` and `yarn.lock`.

Version 2.0 is the exact same version that is in `react-native`.
While there are newer versions that have exciting features, the
first step is to simply start using this new package.