refactor, messageActionSheet: Rename `Reply` to `Show conversation`.
This option was added long time back with an intension that on
selecting it we will narrow down to that message and focus compose
box. In recent refactor of messageActionSheet, focus functionality got
removed (also it was not implemented in the solid form).

Now with this renaming it looks more concret and logical.

part of #3218