Ensure a notification is not narrowed to more than once
Checking for existence of the `config.startup.notification` so we
narrow to it is not always ideal, as we might recheck this again
if we switch accounts, login & relogin and maybe introducing other
cases in future.

We use that config key both to narrow to a notification and to
optionally display diagnostics information. So we can not just
reset it to `undefined` after narrowing.

Instead, we introduce a new flag `shouldNarrowToNotification` that
is initially 'false', then is set to `true` on a new notification.