android, composeBox: Fix placeholder text padding irregularity.
Few observations:

* Only reproducible when compose box is shown after message loading
(GIF like loader) after narrowing on Android. Can't reproduce on
narrows whose messages are cached and immediately visible after
narrowing.

* If `this.getCanSelectTopic()` condition is removed from topic input
and `multiline` is added to `Input`, it start behaving like `Message`
input.

So from this it can be inferred that:

* It is a `multiline` input issue, which occurs when its parent first
got rendered with `display: none`. Normally it is not reproducible on
topic input because at the time of first render (when style is
`display: none`), topic input is not rendered (because
`this.getCanSelectTopic()` is false).

So to fix this issue, this commit replace `display: none` with
`height:0`, without breaking any functionality.

Fixes: #3295