Add `processMeMessage` function
This function checks if the message is of the type `/me ...` and
converts it into format appropriate for rendering. This might have
beeen better implemented server-side.

While the check seems a bit simplistic and possibly error-prone, it
is the exact same logic that the web uses.

We will do a more simple styling that imitates Slack so the producedg
markup is good enough.