messageLinkPress: if narrow link can't be parsed open in browser
Currently `getNarrowFromLink` handles links that it can't parse in
an inconsistent way. It always tries to return a valid narrow but
that is suboptimal.

Here we introduce the first case where if the link type is unknown
we return a 'null' (but more cases to follow)  In these cases,
do the default behavior of opening the link, which is vastly superior
to opening 'Home' narrow or doing nothing.