Refactor and cleanup `getNarrowFromLink` tests
Makes several changes to the `getNarrowFromLink` tests but does not
add any new tests, nor does change any of tested criteria.

 * Some diplicate tested conditions were removed
 * Few tests were split as they were testing different behaviors
 * One assert per test achieved (should almost always be followed)
 * Use `groupNarrow` and `specialNarrow` instead of an array variable
 * Organize the tests consistenly in the same way

Note: Most tests are organized in the AAA (arrange, act, assert)
pattern but this rule is followed somewhat informally (we do not
add comments as some beginner examples do) but is still followed
in the order we arrange the lines of code.