getNarrowFromLink: Parse stream IDs in links
Recognize the new URL format that includes stream IDs and use that
instead of the name to determine the stream name to detect the
stream name. This makes sure the link remains valid through a
stream rename.

This code also supports the old URL format that does not include
a stream ID in the URL.