android, unicode urls: Fix correct url not passed to custom tabs.
Before `getFullUrl` was adding `realm` even if unicode url string
starts with `http`. This issue is completely wired. It can not be
reproducible when remote debugger is on, on Android, but can
otherwise.

Replacing `startsWith` with `indexOf` doesn't make any logically
change, but have internal implementation difference. Thus this got
fixed.