notif: Only send push token to server if it doesn't already know.
One effect of this should be to mostly resolve #2128, which is the
silly issue where the app would tell the server its token over and
over again on switching accounts, with a larger and larger burst of
such requests on each switch.

We're still accumulating a listener on every account switch, which is
the real root cause of that issue; but once the server acknowledges
receiving the token (a fact we save in persistent storage), those
listeners will do nothing unless the token changes, rather than each
making a repetitive request to the server.

We'll fix the root cause, too, shortly.