notif: Cut `config.startup` entirely.
It never made a lot of sense to put this kind of mutable user data
into a "config" object, alongside such things as what kind of logging
to do, or how many messages to fetch from the server at a time.

Now we've just cut out the logic that went through this pathway.  That
saves us the trouble of refactoring it into something more reasonable;
instead, just cut it out.