webview, openLink: Get the AWS res file path and open it in browser.
Auth needs to be passed to access user_upload, or we can also access it
via appending `api_key` to the url. So append the `api_key` to the url,
and then get the redirected url (which is of AWS). And finally open it
in browser, so that user don't have to login to realm in browser.

So the external browser only sees the "expires in 1 minute" S3 URL for
accessing the file. And the `api_key` is also not exposed in the url via
accessing the file in browser.