PresenceStatusIndicator: use `email` prop instead of `presence`
The previous iteration of `PresenceStatusIndicator` used a single
user's `presence` entry to determine what to render. That was not
clearly better than having this component extract the data on its
own from the state.

Now though, we need to also have the `userStatus` data and manually
passing this becomes too much boilerplate.

This change only replaces our use of pre-extracted `presence` data
with getting an `email` of a user instead and doing the extraction
inside the component.

A lot of container components did change, but the change is
straightforward, covered by tests and simplifies our code.

Removed `exraData` prop in `UserList` as not needed (and misspelled)