PresenceStatusIndicator: use 'user status' when rendering
Now `connect`s to the `userStatus` state and if `away` for the
given user exists and is set to `true`, we override the calculated
value and make it `offline` instead.

It is unfortunate that there is a (temporary) discrepancy between
our `presence` and `userStatus` key used for mapping (`email` vs
`user_id`) thus needing the `usersByEmail` state too.

We should be moving `presence` to 'id to presence mapping' and then
we would replace the `email` prop with `userId` prop. Later.