data: Stop searching through all users with `.find`.
Ouch, this is an expensive scan to do routinely when we just want
one user's information -- especially in a large org, with thousands
of users.  Use a decent data structure instead.

There are still some other places where we search through e.g. all
human users; we'll get those separately.

This is a first step on #3339, fixing 5 of the (at least) 9 searches
we found there that go through the list of users.