data: Eliminate getUserByEmail, with its `users.find`.
Just like a few commits ago with `getAllUsers` and its call sites
which would go and scan through all users linearly to find a single
user's data: this is what nontrivial data structures are for.  Use one.

This addresses one more of the `users` searches listed in #3339.
A few more remain.

The use of NULL_USER here is of the kind that is always bad news and
should be eliminated.  For this commit, though, preserve the behavior
unchanged.