Add `UserStatusScreen`
Create a new screen allowing the user to set a status text.
The screen is stand-alone as it dispatches the needed Redux
action to change the value.

Possible presets are given for a one-click entry. Currently
they are taken verbatim from another product (rhymes with Flack)

Values are imported from a separate file as Prettier does not
'like' unicode and messes up the formatting.