testHelpers: Make types strict-local.
The use of `mixed` means no assumptions about this bit are permitted.
If the two test files that use this helper are to be typed in the
future, we'll need to relax that... perhaps by making this actually a
valid value of the type it's really pretending to be, which is a
NavigationState.  No sense doing that work now, though.