userStatusSelectors: Create three user-status selectors
* getSelfUserStatus

Extract the user status object for the logged in user.
If no away status and status text have been set we do not have any data thus `undefined`.

* getSelfUserAwayStatus

Returns the effective `away` status of the logged in user.
It is `true` if explicitly set to that value.
If no value is set we consider it `false`.

* getSelfUserStatusText

Returns the effective `status text` value of the logged in user.
If it is set we get as result that value.
If no value is set we get a valid but empty string.