travis: Dedup runs of the Android build.
Recently, in
  e8e67e8ff tools/test: Add Android build, when applicable.
we added the Android build to `tools/test`, for convenience
in development.

A side effect was that we started running it there in CI; but
we already had a separate line in our Travis config to run it.
So recent builds in Travis have run the Android build twice.
(Fortunately, at least, the second build was short, because
its work was already done.)

Take out the redundant call; `tools/test` can handle it.