android notif: Avoid the nullable `getCurrentActivity` on initialize.
I introduced this path in f507b025c a couple of weeks ago, as part of
upgrading us to FCM.  Things have worked fine using debug builds since
then... but just now, after upgrading an install of the release build,
I found it crashed on startup.  Yikes!

The cause was a NullPointerException -- in fact, precisely one
that Android Studio had been warning about.  (That's what comes of
ignoring a warning!)  Specifically, RN's `getCurrentActivity` method
can return null.

I'm still not quite sure what the story is in that situation...
but in any case we were only using that as the first in a long chain
of methods to get at the `ReactContext`; we don't really need the
`Activity`, or even the `Application`.  And from this call site,
we have a much more direct way to get the `ReactContext`:
`getReactApplicationContext`.  So use that.

(This contrasts with `InstanceIDListenerService`, and with the pre-FCM
`RegistrationIntentService` which is where this code was originally
written -- in those contexts we unconditionally have an `Application`,
and don't have a nicer way that I can see to get the `ReactContext`.)