AccountDetails: Redesign to fit screens better.
Long time ago we made a design decision to show the user's avatar
as wide as the screen (so quite large). While this makes it
attractive to look at, the size is often a bit too large.

Phones with small screens or phones with wide screens (which makes
the height of the avatar a bigger percentage of the total screen
height) leave too little space for other UI elements.

Now that we are showing a bit more, and especially in the upcoming
`ProfileCard`, it makes sense to 'tame' this UI a bit.

This makes the avatar a fixed size of 200x200 leaving enough space
below it even on smaller screens (tested on iPhone 5s).