drafts: Cleanup state when switching accounts
We previously did not reset the state on switching accounts but
only on logout.

This fix has the added benefit that a user can cleanup the drafts
by switching accounts (but this is also a aside effect that we will
not persist the drafts if the user switches accounts and gets back).