ComponentList: Fix a layout issue adding extra space at bottom.
Fixes a bug that adds an extra space at the bottom of the list of
components if the last component is `null`. While that seems it
would occur rarely, if ever, this happens in `AccountDetails` if
the `timezone` prop is not set (we render a `null`).

We add spacing between elements by applying a `marginBottom` to all
but the last component. If the last component (or components) are
`null` this logic adds an extra space that misaligns the view,
and in cases where the screen real-estate is at a premium that is
not a minor problem.

The fix simplifies and makes the code more clear. We first filter
out any `null`s and `undefined`s and in a separate step apply the
styles, using the same logic, which is now correct.