Update architecture.md
Issue #3307 
I think having an index of sorts of what is inside the source directory would help tremendously for beginners to feel more comfortable with the codebase. Although some folders are self-explanatory, new people might not have the same coding background or might not understand as easily, therefore, I am going through all the folders and trying to get an idea of the function they perform( it will help me also :smile: ). This is the first time through and so I wanted to make a pr to ask if any changes are required in the way I have written the description or if something needs to be added/deleted/updated.