avatar: Specify type of `avatar_url` properties more precisely.
The `avatar_url` value coming from the API and used throughout our
codebase can be either a string, or `null`, but never `undefined`.

This makes the `?string` type not precise enough, as it allows
`undefined` values.  This changes each usage of the `avatar_url` value
to the more precise `string | null`.