android notif: Parse zulip_message_id at crunchy shell.
This is a typical example of the "crunchy shell" pattern at work.
Previously, if for some reason we received an FCM message with no
`zulip_message_id`, or one that didn't parse as a decimal integer,
the app would simply crash.

Now we'll instead ignore that FCM message, and log a warning to the
system log for possible use in debugging.