android notif: Parse sender_avatar_url at crunchy shell.
Since at least server v1.6.0, we've never sent a `base_url` --
so the getter `getBaseURL` has never returned anything but `null`.

So, first: hard-wire that to `null`, and constant-fold it through
several layers where a `baseURL` value was passed.

Doing so, we learn in `addHost` that if we'd ever had a
`sender_avatar_url` which didn't start with `http`, we would have
blown up with a NullPointerException on trying to use `baseURL`
to fix it.  This seems to be meant as a way of testing for a
relative URL reference, vs. a real absolute URL (though it's an
awfully buggy one.)

So, pull that check up into the "crunchy shell" of parsing, and
handle it appropriately.  Also replace the check's logic with a
non-buggy version: check that the URL parses as a real, absolute URL.

A question this raises is: how likely was it that we'd get one of
these crash-causing FCM messages (with `sender_avatar_url` not
starting with `http`) in the wild?  In the server Git history, it
looks like server v1.6.0 could send `sender_avatar_url` as a
relative URL if it's a local upload, or if there's no upload and
ENABLE_GRAVATAR is false.  But commit 1.7.0~3388, released Oct
2017, fixed that; since then it's always absolute.