android notif: Move top level of parsing inside new design.
The first step in parsing an FCM message is to look at the `event`
field, which the server sets to identify the type of message.

We already handle this step in "crunchy shell" fashion; but move it
inside the same framework as we've made for the parsing steps that
come after it, in order to encapsulate all our parsing there.