notif: Cut unused fields from type definition; add jsdoc.
Many of these fields aren't always there: e.g. `google.ttl` isn't
anything we ever set (just as one would hope), and perhaps FCM sets it
or GCM used to, but surely APNs doesn't.

Perhaps worse, because more tempting for us to actually try to use:
the Zulip server only started sending `sender_id` in 1.8.0-rc1~1860
(and 1.8.0-rc1~1950 for Android devices)... which were released
less than a year ago, in 2018-04.  We could use them, but it'd be
important to properly handle their absence -- in particular not crash.

Some of these are fields it'd be good for us to start using;
we'll want to do that behind a "crunchy shell", as the `sender_id`
example illustrates.  For now, just remove the false advertisement.