titlePrivate: Apply `flexDireaction: 'row'` to view inside Touchable.
`Touchable` can only contain one view inside it, so applying this row
style to it doesn't make any sense. So layout issue was introduced in
8f99a62, where we replaced `View` with `Touchable`.

Now create a new view inside it and apply style to it. Other styles of
`navWrapper` are now taken care by `navContent` (which got added in
previous commit).