ChatNavBar: Eliminate the thin TitleNavButtons module.
This module did very little work; as a result, the complexity it
handled for us was no more than the complexity of the interface to
use it.  Instead, just have its one caller directly do what it did.

This does leave ChatNavBar dealing with a bit of mess around the case
where no button is needed for the given type of narrow.  That bit of
complexity really belongs with `getInfoButtonFromNarrow` etc., which
already have the responsibility of working out what to display for the
given type of narrow; we'll move it there in the next commit.