Touchable: Make interface work the same between platforms; add doc.
The upstream components we choose between to implement our Touchable,
namely TouchableNativeFeedback and TouchableHighlight, turn out to
differ in a subtle but important aspect of their interface:

 * TouchableNativeFeedback returns just its child, with extra props
   to make it a touchable area.

 * TouchableHighlight makes its own wrapper View to be the touchable
   area around the child.  The `style` prop is passed to this wrapper.

Our Touchable code was adding another wrapper View in both cases,
between its `children` and the upstream Touchable.  The result
looked like:

  on Android, with TouchableNativeFeedback:

     (touchable area,)
     (with `style`   ) -> (children)

  on iOS, with TouchableHighlight:

     (touchable area,)
     (with `style`   ) -> (wrapper) -> (children)

This meant that in the iOS case, certain `style` properties like
`flexDirection` and `justifyContent: 'space-between'`, which
operate on a node's children, wouldn't work as expected if passed
to Touchable -- they'd control only the wrapper, and not the
actual `children` passed to Touchable by its caller.

This resulted in bugs like the one fixed recently in b78b49d62 in
TitlePrivate.  The confusion can also be seen in places like
ZulipButton and AccountItem that pass to Touchable `style` values
containing these properties.

To clear things up, we give both cases a common interface:

     (touchable area,)
     (with `style`   ) -> (child)

In the TouchableHighlight/iOS case, this requires us to insist that
`children` actually be exactly one child.  Turns out we already do
this at every call site of Touchable, and have for a long time (with
the brief exception of the period up to b78b49d62 after the bug it
fixed was introduced) -- no coincidence, as the existing behavior
makes it nearly impossible to use the "multiple children" feature of
our existing interface without having a layout bug on iOS.