tools/test: Fix selectivity in Android case, using base commit.
This gives the `android` suite the same nice behavior as `lint`,
`jest`, `prettier`, and `spell`.

Before this change, we would decide to run Gradle by comparing
your current tree with the *tip* of upstream -- so if you had
made no Android-related changes, but there'd been such changes
upstream since the start of your branch, we'd unnecessarily go
and run Gradle.

Instead, compare to the *base* of your branch where it forked
from upstream.