android notif: Parse `pm_users` in the "crunchy shell".
When we pass this on to our JS code after the user opens a
notification, that code will go and try to parse it.  Ensure up front,
here in our crunchy shell, that it will actually parse successfully.

A further good step would be to actually pass JS a list of ints,
rather than re-encode as a string.  That can be separate, though.