messageHeaderAsHtml: Modify topic header to include date.
Many times when people are scrolling through messages they need to see 
what date it is. This modification is made for this use-case. Users can 
simply see the header to determine the date a particular message was sent.

This is implemented using some basic `css` and `js`. When writing the css 
selector for the header of the topic, there was an issue arising where the 
sticky header (because a small part of it gets cut out) was showing 
misalignment. This is the reason why `margin-top: 0.2em` was added to 
`.topic-header` selector.

Tests have been conducted, and behaviour is checked for both Android and 
iOS, light and dark mode.

Fixes #1336