webview js: Remove a "temporary" bug in click handler.
Commit dd03939cb from yesterday adds an explicit exception in the case
of certain programming errors (where previously we would have failed
in more complex and unpredictable ways further down the line), and I
wanted to confirm how an exception there was handled... so in order to
test that, I artificially introduced an error in one such spot.

Then I committed the change with that error included.  Oops. :-/

Take the error back out.