react: Use explicit boolean checks in `check && <Foo .../>` pattern.
Flow 0.92 complains about these with errors like this:

  Cannot return function because:
   • Either inexact React.ComponentType [1] is incompatible with exact
     React.Element [2].
   • Or React.ComponentType [1] is incompatible with React.Portal [3].
   • Or property @@iterator is missing in React.ComponentType [1] but
     exists in $Iterable [4].
  [...]
       48│   return SpecificButton && <SpecificButton {...props} />;

where `SpecificButton`, the thing left of the `&&`, is highlighted,
and it has type `React.ComponentType | null`.

Seems like it's concerned that something in ComponentType could be
falsy so it ends up getting returned.

Well, certainly that would be bad if it did somehow happen.
Let's just be explicit about treating this bit as purely a boolean.