flow: Use better types for `navigation` properties
Use the more precise typing for `navigation` which is:
`navigation: NavigationScreenProp<NavigationStateRoute>`

This would actually catch errors if we did have any, as now the
value is properly typed.

We do not lose any typing from the:
```
state: {
  params: {
    narrow: Narrow,
  },
},
```
As the `<*>` would allow anything as alternative.

I experimented with typing that did not work, similar to:
```
export type NavigationParams<T> = NavigationScreenProp<{
  ...NavigationStateRoute,
  params: T,
}>;
```