msglist header: Modify topic header to include date in stream narrow.
Many times when people are scrolling through messages they need to
see what date it is. This modification is made for this use-case.
Users can simply see the header to determine the date a particular
message was sent.

Header is converted into a flexbox and js for timestamp is added.
The date is human readable, and header is truncated if the topic
name is very large. This commit only adds the feature for narrows
where headers already exist; subsequent commits add headers to the
rest of the narrows.

The headers are changed in architecture slightly, and can now be
described as modular. To understand this, consider `topic-header`,
and `stream-header`, and `private-header`. These three types are
essentially modules which represent different styles the headers can
be displayed in. In accordance with this change the stream header is
now a module plugged into the normal topic header which basically
displays the stream name `# stream-name` before the topic name.

This commit introduces changes which slightly disrupt visual in pm
headers (which appear only in special narrows) because of the slight
semantic change in how padding is used in the header.  Specifically,
the horizontal padding in the header is lost:
  https://user-images.githubusercontent.com/33200024/56861714-480e1d80-69c1-11e9-99e7-350f8acbff94.png
Because this bug is small, we'll tolerate it for the moment and fix
in a subsequent commit.