connect: Use 'navigation.getParam()' instead of 'navigation.state.params'
This change preserves the behavior and is almost a pure refactor.

Using `getParam` is the preferred way of accessing the navigation's state
parameters and has some advantages, as documented here:
https://reactnavigation.org/docs/en/params.html

`navigation.state.params` can be `null` if no parameters are passed.
While we do not do that in our code, the types, nor our code is
acknowledging that possibility.

`getParam` will always execute and we can optionally provide a
default value when none exists. Sometimes, not giving a default
value is good enough.

Prime example of how this approach simplifies our code (if written
correctly) is the `RealmScreen.js` file:
 * we don't actually need to verify if `navigation` is defined,
   it always is, when the component is used as a navigation screen
 * we don't need to check if `navigation.state.param` is defined,
   as `getParam` will do that for us
 * `getParam` also provides us the default value