MessageList: Fix completely bogus `||` on optional boolean.
This conditional has literally never worked at all since it was
introduced in d2273c612.  The caller passes in `false`, attempting to
short-circuit the call to the selector... but of course `false` is
just as falsy (!) as `undefined`, so it has 100% exactly the same
effect as not passing in the prop at all, and we go ahead and get the
value from the selector.

The reason the UI still looks right is that, basically by accident,
the selector always returns `false` for a search narrow: in particular
we're never "fetching" in a search narrow.

Found by marking this file strict-local, and seeing what errors
that produced.