msglist header: Modify pm/group headers to show date - special narrows.
The previous commit changed only the headers in stream narrow. This
commit adds the same for the private/group headers in special
narrows. The implementation for private narrows was straightforward
for special narrows.

Renamed the `topic-date` and the `topic-text` to `header-date` and
`header-text` to reflect their actual usage, since they are used
everywhere, and not only in representing topics.