nav: Broaden type fixme to all screens in AppNavigator.
As mentioned in the commit message in a36814e80 which originally
added these fixmes:

  In fact we'd need fixmes on more of them... if the react-redux
  libdef didn't defeat type-checking at `connect` calls.

We're about to get a version of that libdef (plus our own wrapper)
that *doesn't* defeat type-checking at `connect`... so a bunch more
of these will need fixmes.  Just write one fixme to cover them all.