connect: Convert remaining error sites to use "fixme" version.
These are the call sites that show type errors when we
switch to the upcoming improved `connect` type that lets
the type-checker work.

Done mechanically.  While on a version where `connect` had
the improved type, ran the following command:

  $ node_modules/.bin/flow status --json \
    | jq -r '
        .errors | map(.message[0].loc.source)[] ' \
    | sort -u \
    | xargs -r perl -i -0pe '
        s/\bconnect\b/connectFlowFixMe/g '

where the first few steps of the pipeline identify exactly the
files where Flow is reporting errors.  Then reviewed the diff
and fixed up by hand the one place (in RealmScreen) where the
word "connect" appeared as a false positive.