types: Again make toplevel-declared types exact, where no errors.
Similar to 703739338 a few months ago: done for the same reasons,
and in the exact same way.

The remaining cases that cause errors are mostly examples of #3451,
so we can convert them to exact object types as we fix that.

And then there are a handful of types that are genuinely meant to
be inexact object types.  When we upgrade soon to Flow v0.84 (as
part of #3399), we can make those explicit; tracking that as #3452.