webview js: Use `clearTimeout` to clarify long-press logic.
The current implementation of how long-press events are handled is a
little convoluted, and it's hard to clearly understand what the code
is doing and why certain pieces of code are present.  For example, in
the `touchend` listener, where `lastTouchEventTimestamp` is set to be
`Date.now()`, there is no intuitive reason why it is determined to be
so while in the `touchmove` and other listeners it's set to 0 instead.

This commit removes the variable `lastTouchEventTimestamp` entirely,
and instead uses `clearTimeout(someTimeout)` to clear the timeout
set for triggering the long-press handler.

This is a pure refactor -- it shouldn't make any change to the visible
behavior of the code.  To explain why that's true: The effect of
setting `lastTouchEventTimestamp` was to prevent an upcoming call to
`handleLongPress` from having any effect, through the status-checking
code at the top of that function.  Instead, in each case we use
`clearTimeout` to prevent the call from happening at all.