stream list: Prevent creating or editing streams with the same name.
Display error message and disable submit button when user tries to
create or edit a stream with the same name (case-insensitive) as an
already existing one. The error is not shown when name field is empty
or it has not been modified.

The error message replaces "required" field and the input underline
color is set to red (on Android).