MessageList: Better propagate read-only flag on some types.
We've long had incomplete type-checking on this component, because of
the use of `connectActionSheet` which lacks a useful libdef.

With our spiffy new well-typed `connect`, and by experimenting with
adding a `SelectorProps` type for the `mapStateToProps` return value
in order to bypass the ill-typed `connectActionSheet`, it turns out
there are in fact some type errors that that has been hiding!

For this commit, fix one relatively mild class of them -- the
selectors return `$ReadOnlyArray`, but various lower layers have been
taking plain array types, even though in fact they do treat the data
as read-only.  Just mark that in their types.