lightbox: Avoid attempting the lightbox on an outbox message.
I'm not 100% sure this is even reachable in the case of an
outbox message; but the upcoming improved type-checking on
`MessageList` (via adding a SelectorProps to bypass the
ill-typed `connectWithActionSheet`) shows that as far as the
types are concerned, it might well be reachable.  And it
doesn't seem likely to go well if we proceed -- the image
may not be on the server for the lightbox to show.

So, add a check to ensure that we don't press ahead and show
a broken lightbox.