webview js: Make human readable timestamp react to scroll.
Why this was required :-
The previous implementation of timestamp in header did not react to
changes in dates. The dates remained the same whether or not a
timerow passed by in the message list. This implementation fixes
that problem and makes the timestamp reactive to scrolling events.

Implementation :-
Every message in the webview now includes another attribute which
shows what date (human readable) it was sent.
In the webview js, after everytime the scroll event fires the first
visible message is calculated, and the closest header is walked
to. The human readable date is extracted from the attribute
introduced in message, and the content of the timestamp in the
header is changed to suit that.

Things to be considered :-
This might induce some performance issues. Optimisation seems to be
unnecessary for just one change in the DOM every scroll event,
however this is still something to watch for in case users report
slowdown in performance. This also has been tested for both devices
on simulators and there were no stutters on them.

Fixes #1336.