webview: Simplify 'scroll-bottom' button
The initial implementation of the 'scroll to bottom' button had
some extra 'scroll-bottom' classes applied in order to simplify
the 'onClick' checks when the target is any of its children.

I was recently reminded of the 'pointer-events' css property.
Setting it to 'none' allows us to have elements that do not cause
touch/click events.

This gives an explicit and unique class to the `a` tag and removes
the extra classes. The new css ensures that any child of `a` will
not cause our 'click' event handler to fire.