msglist header: Refactor which-header-style logic slightly further.
This adds a bit of additional organization to the handy logic
added in the parent commit, grouping similar types of narrows to
make it a little easier to see at a glance which of them get what
style of header.

Also, for indicating an absent header, replace `null` with an explicit
string value, making it symmetric with the other styles.  And write
down the type of this enum, because on its own Flow doesn't infer that
this is an enum and just sees it as `string`.