UnreadCards: Simplify implementation and types
Using `...restProps` should be kept to a minimum and used only in
rare places that it makes sense to. We follow this practice and
use it very sparingly.

The only place that it doesn't make sense to use it is `UnreadCards`
where we know all the remaining props this refers to, currently,
for no good purpose.

This change simplifies our code and makes it more explicit what
property we rely on (just one!)

We can now also simplify the Flow types.