PmConversationList: Get 'dispatch' injected by 'connect'
In the past, there was a good intentioned practice to try and
separate 'store-aware' from 'store-unaware' components. The latter
wouldn't use `connect` and would get all their props via their
parent, they would not be aware of Redux.

Some components had to call a Redux action though, so we did pass
the `dispatch` function to them to use.

We have moved away from this approach, now following the opposite
practice - every component that needs store data uses `connect`
and gets it locally, not relying on their parent component.

This is working well. We should keep doing it.

There is only one instance where we still pass 'dispatch' manually,
so do add 'connect' to 'PmConversationList' and stop passing it
from the parent component 'PmConversationsCard'.