flow: Fix 'accounts' prop type in 'AccountList'
This property is coming from the `getAccountStatuses` selector
which returns a '$ReadOnlyArray' type. This is great (and correct)
but is incompatible with the basic, writeable version.

Upcoming commit that types the `connect` call that provides this
property to the component will result in a type error, so we are
fixing this early.

Note: it would be great to be able to mark in a more seamless way
all properties coming from the selectors and all component props
as 'read-only'. It is bad practice that may produce bugs. While
we do not intentionally mutate any of these values, catching
instances where we do would be valuable.