msglist header: Modify pm/group headers to show date - pm/group narrows.
Adding headers to PMs/groups was a little tricky.

In private narrows we don't want the names of the individuals in the
header because they are already present in the AppBar. The changes
in `messageHeaderAsHtml` reflect this. There is an empty element
introduced which acts as a buffer to enable the flexbox to do it's
thing. Also edited `renderMessages` to actually add the headers to
the `renderedMessages` pm/group sections in the first place.