android, notif: Change structure of the ConversationMap.
Make it LinkedHashMap of Identity and ByConversationMap, so that
messages are segregated according to each Identity. This will help in
showing seperate notification for each Identity.

If fcm message doesn't contain server info (Identity,this  means server is
old), then Identity  will be null and will fallback to previous
implementation i.e a single notification for all these servers.

Now each Identity have it's own notification. But we have hardcoded
single notification id, so it is not possible to have multiple
notification at the same time. Now next step will be to get different
notification id for each case.