selectors: Reserve "from foo" for when the data comes *from* foo.
Many of our selectors return some data that's specifically about a
particular user, stream, etc. identified by the selector's argument.
For these, the naming pattern we usually use is "getBarForFoo" where
the "foo" is the identifier passed as an argument -- out of all the
bars we have, the selector gets the particular one that is *for* the
given foo.

Some other functions (e.g., `getAvatarFromUser`) are passed an actual
blob of data (in this example, a `User`) and the data they return is
actually taken *from* that blob of data -- not from somewhere else
like the Redux state.  For these, the naming pattern is "from foo"
rather than "for foo".

We had a handful of selectors that were named "from foo" when what
they actually do is get some data "for foo".  Fix those.

Renames carried out mechanically, like so:

  $ perl -i -0pe 's/getStream\KFrom/For/g' src/**/*.js
  $ perl -i -0pe 's/getSubscription\KFrom/For/g' src/**/*.js
  $ perl -i -0pe 's/getAccountDetailsUser\KFrom/For/g' src/**/*.js