api: Add custom errors classes 'ApiError' and 'EventQueueError'
We are currently hackily injecting some API data to a standard
Error object. This works fine but creates issues with Flow, as
it rightly complains of misuse.

The right approach is a custom class, descendent of the Error
object and enhanced with the needed properties. This used to
be tricky because of the Babel / React Native build pipeline
but improvements in the past year have made this possible.

This creates a base 'BaseApiError' custom error class.
Then we create two other classes that extend it: 'ApiError' and
'EventQueueError', mostly for typing purposes.