events: Log out if polling for events gets a 401 response.
Fixes #3253.

When the user rotates their API key, or if they or their realm simply
disappear or are disabled, all our requests with those credentials
will start failing with HTTP status code 401.

In the previous commit, we started detecting that case where we try to
register a new event queue; in this commit, we do the same when
polling an existing event queue.  Now we'll detect it whether or not
the app is continuously running and trying to use the newly-invalid
credentials.