brief-messages: Add timestamp.
There have been multiple complaints about differing timestamps not
being available for messages sent at differing times by the same
user in succession. This commit aims to solve that problem by adding
a timestamp which comes right after a brief message (on the same
line, so as to conserve space). This timestamp uses the
`float:right` property, and clears itself from timestamps appearing
in further messages.

There is a dummy element added in the `widget` builder in
`messageAsHtml` in order to properly align the timestamp which goes
along with it. Since this element is temporary and is a dummy itself
anyways, it will not be a problem for code quality or
understandibility since it is going to be replace soon.