widget: Put the CSS near the other message CSS.
The chronological order of features added isn't a good way to
organize this CSS file, and indeed we've largely done better than
that.  Instead of putting this new bit of CSS for a special category
of messages at the very end -- after the styles for things like the
`#scroll-bottom` button and JS error banners, which live outside the
sequence of messages -- put it just after all the other styles for
messages.