compose menu: Refactor function names.
`uploadImage` has been changed to the more appropriate `uploadFile`
since it uploads every file and not only images. Appropriately, in
`fetchActions.js`, the function calling `uploadFile` has been
replaced by just the default export of `api` (convenience added in
b8010de in an effort to solve a similar problem encountered before).

Also changed `handleImageUpload` to `handleImagePicker` for that is
the more accurate version of what the function is doing.