webview: Add auth to local documents
Fixes #3487
Fixes #3303 (but does not implement the exact suggestion)

Just as we do with images, check if a link is local and is poining
to `/user_uploads/`. If so, adds the `api_key` value to the url
to enable opening it in Chrome Custom Tab / Safari View.

The file can then be downloaded via the oveflow menu (three-dots)
on Android. On iOS, similar actions can be taken via the 'Share'
button available at the bottom of the screen.