webview css: Use `rem` instead of `em` as a length unit.
We use 'em' as unit to make an element scale with the
set font-size. That is very useful, as we can mix these with
the conventional 'px' to ensure other elements do not scale.

This commit changes all our 'em' units to 'rem'.  'em's are
relative to the parent element's font-size while 'rem's are based
on "root em" - the font size of `<html>` - which is exactly what
we want.

This change is strictly to make the styling more robust and does
not change any elements' size.  While the two units are not
equivalent in general, in our use they mostly are: they differ only
when the parent has a `font-size` different from the root.  We set
`font-size` only on the root element `<html>` and on various small
elements near the leaves of the DOM tree; of these, the only ones
containing child elements with their own styling are `.message-tags`.
So to keep this change a pure refactor, on the `.message-tag` child
elements we adjust the numbers to compensate for the change in units.

Further reading on 'em' / 'rem':
  https://j.eremy.net/confused-about-rem-and-em/
  https://zellwk.com/blog/rem-vs-em/

[greg: fixed equivalence argument, added `.message-tag` correction]