timerow: Refactor - Change inaccurate names.
(references parent commit)
The name `timerow` was not very correct, and its child introduced in
the parent commit isn't very compliant either. This is because even
if the name was a befitting the function before, since timerows were
a divider of sorts (they still are), the design has changed to the
`pill` type, and that description is more accurate now.

This commit refactors the names of `timerow-content` and
`timerow-persistent` to more appropriate names.