webview css: Set `font-size` on `html` instead of on `body`.
We use `em` as a unit to make some lengths scale with the font size.
We'd like to switch all of these uses to `rem` ("root em"), which
refers directly to the font size of the `<html>` root element, making
the relationship to `px` much easier to robustly reason about.  See
discussion in docs/background/webview.md for more background.

As a preparatory step, in this commit we move where we set the
`font-size` property, to do so on the `<html>` element rather than
`<body>`.  This is a pure refactor, having no effect on the layout,
because either way it affects the entire document (as we don't yet
use `rem` anywhere.)

[greg: split out from next commit; wrote commit message]