docs/release: Move manual testing to post-alpha.
Recently I've usually been doing things in this order.  It lowers the
barrier to getting an alpha out, and it facilitates a more faithful
end-to-end test: the exact build uploaded, on my physical device.

This does mean a greater risk of having to immediately cut a new
alpha.  But we're pretty good at not letting issues slip into master
in the first place that this manual testing would catch.  Also, the
new script `tools/ios` from earlier this year has made the build and
upload significantly less painful to do.