message list: Fix misalignment between messages with/without avatar.
Fixes #3508.

The apparent spacing of the left margin is different between a
`message-full` (one where we show the sender's name and avatar) and a
`message-brief` (where we don't).

This was introduced recently by commit 3dd5fb4de "Make avatars bigger,
and not scale with font".  To fix it, we partially revert that change,
making the avatars smaller again.

It'd still be good to make the avatars bigger; see #3515 for that.

[greg: added new issue #3515, expanded message slightly]