msglist: Let font size on iOS scale with system settings.
Fixes #2559.

Instead of setting a hard-coded font-size of 15px, use the equivalent
percentage to get 15px when 16px is the default, which it normally is.
This way, when a user has chosen larger or smaller text in their
system settings, the text in Zulip can scale appropriately.

Empirically this already worked on Android; perhaps the platform has
some hack to work around applications doing exactly the kind of broken
thing we were doing.  So this change has an effect only on iOS.