css: Remove extra spacing between two message-fulls.
When two message-full objects come adjacent to one another, the
distance between them becomes 3rem, which is not desirable at all.

This commit removes some of the code produced in the parent commit
which was written in order to keep things consistent and not change
the design of the webview too much in one commit. The removed code
used to add extra padding to make the total 3rem.