msglist css: Convert more layout to not grow with font size.
There are still other spots that need this treatment, but these
were some of the easy ones.  Like elsewhere, this means that when
the user has expanded specifically their font size, we don't have
our padding and other non-text areas grow to encroach on the text
content, which in that situation will already be getting cramped.