mentions: Fix highlight on mention extending disproportionately.
Mentions highlighting extends to the bottom and not at all at the
top. This commit fixes that, through some CSS which might be
confusing to a reader who is already familiar with the current
codebase.

Because of the changes made in this commit there are a few cases
where the design is not exactly the same as before this commit. One,
which `.message + .message-full` rule temporarily fixes, is where a
message-full directly follows any other message. (The bug' was that
there appears a loss of 24px of `padding-top`.)

Second, and third, are introduced at the extremes of the message
list. At the very top of the message list some extra padding exists
because the topmost element will always be a `timerow`. And, the
bottom most element will have a reduced padding -- because of the
fact that we're changing the bottom padding of all message elements
by 8px (16px -> 8px), and the bottom most element (which has
padding) will always be a messages.

Fixes #3513.