messageActionSheet: Add option to copy link of message to clipboard.
There are use cases where a person wants to share a message with others
who he knows have access to a particular stream. To resolve this, in this
commit, another option is added in the action sheet which shows up when
 a message is long pressed, which  copies the link of the message to the
Clipboard.

I have edited the buttons to say `Copy message content` for copying their
content to the clipboard, and `Copy message link` to copy links which lead
to the message. This commit uses the function introduced in its
parent in order to find the link of the message.

Fixes #2623