api: Add `recent_private_conversations` key in `register` endpoint.
The server added in zulip/zulip#11944 basic support for fetching the
most recent private conversations that a user was involved in.  The
endpoint returns the private conversations a user has had within the
past 1000 private messages they have sent/received.  The endpoint
returns data in the form of `[{user_ids, max_message_id}]` where
user_ids lists all the people involved in the conversation, and
max_message_id returns the message_id of the most recent message in
the conversation.

This commit adds only the minimum machinery necessary to request this
data from the server at `register` time.

[ray: Split original initial commit into two parts, this being the
  first. Minor changes to Flow types.]