widget poll: Add basic support to show widgets from available data.
Widgets support has been in great demand from users (as it was
implemented in the webapp first, and we haven't supported its
display since support was added to the API). You will also see a
`Poll Widget` marker in the bottom right of the box which shows it.

This commit adds the `poll` widget support, and modularises widgets
so it will be easier to add other widgets if required. It uses the
newly introduced (5046781b5) `getOwnUser` to get the `user_id` of
the current user in order to check it against the `key` provided in
submessages (which are only widgets for now) of type `vote`, to see
if the current user has voted on the `option` in that `poll`.

Styling has been duplicated from and merged with `.reaction`, while
the borders of the `.widget` class has been changed to match the
`.static-timestamp` class.

Fixes #3205.