migration: Add remedial migration after #3553.
Now that #3553 is fixed so that migrations actually work again, let's
ensure that any install which missed a previous migration due to that
bug gets the effect of it now.

Fortunately, because `dropCache` is idempotent, this is pretty simple.
We can drop the `AsyncStorage.removeItem`, too, because we've re-used
the same `messages` key since then.