submessages: Add `vote` ability to `poll` widgets.
Clicking on the button which shows the number of votes given to an
option will now add a vote from your side to it, or remove the vote
in case you had already voted on it.

Implemented by adding `sendSubmessage.js` to the api index, which
sends the vote request to the user's realm using an event handler in
`webviewEventHanders.js`.