account: Clear `realm` state too on logout and login.
All of this data is specific to the particular account or to the
realm/org, so it doesn't make sense to keep any of it on logout.
Drop it completely, just as we do in almost all our reducers.

On login, it's less clear that we need to do anything; but clear
the data then too, for uniformity with our other reducers.  If we
decide to change that behavior, we can do it across the board.