zulipSwitch: Always use switch `value` from parent.
In case user changes value, it will reflect in parent state/store and
updated value will be passed to the zulipSwitch becuase here switch is
always used in contolled way (all caller passes `onValueChange`).

So there is no need of the state in switch component to maintain updated
value.

Fixes #3106.