eslint: Upgrade to ESLint 5, and upgrade related packages.
The simplest version of this would have been to upgrade all
ESLint-related packages to the very latest versions:

  $ yarn upgrade --latest $(tools/deps | grep eslint)
  $ yarn upgrade $(tools/deps | grep eslint)

But those aren't all compatible; ESLint 6 is just under a month old,
and a few packages don't yet support it, resulting in these warnings:

  warning " > eslint-plugin-react-native@3.7.0" has incorrect peer dependency "eslint@^3.17.0 || ^4 || ^5".
  warning " > eslint-config-airbnb@17.1.1" has incorrect peer dependency "eslint@^4.19.1 || ^5.3.0".
  warning " > eslint-config-airbnb-base@13.2.0" has incorrect peer dependency "eslint@^4.19.1 || ^5.3.0".

To keep everything compatible, we take the latest eslint@^5 ; and
then the latest eslint-plugin-flowtype is incompatible with *that*,
so we limit the latter to version ^2 .

The resulting command is

  $ yarn upgrade --latest \
      eslint@^5.16.0 eslint-plugin-flowtype@^2.50.3 \
      $(tools/deps | grep eslint \
        | grep -vx -e eslint -e eslint-plugin-flowtype )

Then a further upgrade, this time without changing package.json:

  $ yarn upgrade $(tools/deps | grep eslint)

I don't really understand why the first command leaves anything
undone for the second to do, but empirically it does.  After the
second we're at a fixed point, though, thankfully -- running it
another time has no effect.