android build: More cleanly make SYSTEM_ALERT_WINDOW conditional.
RN apparently needs this permission in debug, for the dev menu.
But there's no need for it in release builds, and we've long
left it out there.

Now upstream has since facebook/react-native@56fc63036 its own way of
making this permission condition on debug vs. release -- and it's a
cleaner use of Android build mechanisms than what we've had.  So,
thanks upstream!  Follow their approach, which also helps keep the
diff small.