ios: Attempt to configure JavaScriptCore.framework .
This is a best effort at merging facebook/react-native@3341adac4
with our Xcode config; that's the entire change to this file that
appears between our current RN v0.57.8 and the upcoming v0.59.10.

Empirically it seems to work, so that's good.

There's been a lot of divergence in this file, between the template
and our version, that I suspect is mostly accidental.  We might be
better off discarding our Xcode config entirely and rebuilding it from
scratch with a current template.  That's out of scope for this
upgrade, though.

I'm a little puzzled that the app builds and runs just fine both
before and after this change.  It's clear I don't totally understand
what the change is doing.  Well, so be it.