nav: Disable tabs animation, to work around breakage.
With the upcoming upgrade to RN v0.59 (from v0.57), without this
change I found that I could navigate to any tab as usual... but
just once, and then the bottom tabs had no effect, until the app
was restarted.

Similarly, if I used my one main-tab navigation to go to "streams",
then the top tabs there were good for exactly one navigation: I
could navigate from "Subscribed" to "All streams", but not back.

So, that's pretty broken.  The issue appears to be
react-navigation/react-navigation#5713.

This change is a workaround suggested in a comment there (thanks!)
The look isn't awesome, but at least it keeps things functioning.

The issue thread sounds like the issue is fixed in a later
release of react-navigation.  We're on a quite old release, 1.6.1.
We have some partial work toward an upgrade (#2702 and #3502);
so this is one more reason to finish that.