ios: Skip a complication in upstream template, explaining why.
This corresponds to facebook/react-native@a2ef5b85d , one of the
changes in the upstream template from RN v0.57 to RN v0.59.

It's a bit hard to understand what the commit is saying.  The docs
it's referring to help a lot -- they've since been deleted, but can be
found here: facebook/react-native-website@6bf0a02fd

On studying the commit, and the code in RN itself which it interacts
with... I'm pretty sure it actually does nothing useful!  And it makes
this code more complex, and also gives a confusing picture of what the
APIs it's using actually do.  So, instead of applying it, just add a
comment saying why not.

Details on reaching that conclusion follow.

First: the commit message links to two issues it's supposed to fix.
But those are both failures at *build* time, with `main.jsbundle`
missing.  That commit does nothing at build time.  That alone makes it
impossible for it to fix the linked bugs, short of some surprising
magic that would cause it to affect build-time behavior.  No such
magic is called out in the commit message (and there's no discussion
on the PR, except procedural), so I'm tempted to think the author and
reviewer were simply confused here.

If they were, they were only as confused as the docs were before that.

Well, and after reading RCTBundleURLProvider... it does indeed look
like the docs were just confused.  First, in the headerdoc for a
version of this `jsBundleURLForBundleRoot:` method (in
`React/Base/RCTBundleURLProvider.h`):

  /**
   * Returns the jsBundleURL for a given bundle entrypoint and
   * the fallback offline JS bundle if the packager is not running.
   * if resourceName or extension are nil, "main" and "jsbundle" will be
   * used, respectively.
   */

So if the packager isn't running, this code already falls back to
getting the `main.jsbundle` resource -- the same thing this new logic
in the template also does.

Well, but what if a packager is somehow running in release mode?  We
don't want to use it, so it'd be good to skip even looking for it --
which will also save latency.  Is the commit helpful for that reason?

In fact, RCTBundleURLProvider takes care of that already too.  Looking
at the implementation shows that if !RCT_DEV (a synonym of !DEBUG),
the packager unconditionally "isn't running" -- the code to look for
it is `#if`'d out entirely.

So the extra complication in that upstream commit
facebook/react-native@a2ef5b85d can't possibly fix the issues it's
intended to fix; and in fact it has virtually no effect at all.
Therefore, skip it.