narrowsReducer: Fix variable names in two loops.
The accumulator in the first loop certainly isn't a `msg`.  It'd be
quite reasonable to call the accumulator in the second an
`updatedState`... but we also call it a `newState`, there's no need
for two different names for it, and we call the analogous thing
`newState` in the neighboring function.  So go with `newState`.