Make `pmConversationsSelectors-test.js` `flow strict`.
The tests for `pmConversationsSelectors` are now `@flow
strict-local`.

To aid this process, we make use of the 'new' way of making tests,
which is using `src/__tests__/exampleData.js`. A few examples of
such tests include the unit tests at
`src/users/__tests__/userSelectors-test.js` and
`src/webview/html/__tests__/messageHeaderAsHtml-test.js`.

Note that this is just a translation, and not a behavorial change.