tools: Exclude deleted files from lists of changed files.
This feels a little awkward in principle... but in the current ways
we're really using these lists, we pass them to tools that want to
go read the actual files in the tree, so it's fine.

And it turns out that ESLint in particular just barfs when one of the
files it's pointed at doesn't exist -- not even bothering to go check
all the other files.  So this gives it the hand-holding it needs.