compose menu: Handle different APIs differently.
The `react-native-document-picker` and `react-native-image-picker`
libraries are separately authored and maintained, and have chosen
different APIs.  In the parent commit, we took responses from the
former and rearranged them to look like responses from the latter,
in order to pass them to the handler we'd previously written for
the latter.

Instead, factor out just the bit that's actually in common, and
handle the DocumentPicker responses in a way that's natural for
that library's API.

This also saves us from the common Promise API mistake of invoking
`.catch()` on the return value of `.then` when we really only
intend to handle errors from the original Promise.  As written,
this code would invoke `this.handleImagePickerResponse` a second
time if for some reason there was an exception thrown when we
called it the first time.

Also the annotation `error: string` on the `catch` handler's
parameter isn't true; drop it.  (Indeed it it were true,
`DocumentPicker.isCancel` wouldn't work at all.)