Omnibus revert to 25.8.122.
There's a showstopper bug #3588 that was introduced, or triggered,
by our upgrade to RN v0.59.  Debugging over the past few days has
made progress, but it's not yet clear where exactly the bug is or
how to fix it.

So, we're going to make a release that just rolls back to before
the upgrade, as a stopgap while we continue to debug.  This commit
returns the tree to what it was at 25.8.122, the (beta-only) last
release before we made that upgrade in 26.0.123.

Luckily the bug only appears on iOS, so we only need to make the
stopgap release there -- lucky because RN v0.59 was the first
version to include 64-bit libraries for Android.  Those are now
mandatory in the Play Store, so a comparable rollback there would
be much more complex.

Revert made with:

  $ git revert -n 25.8.122..
  $ git commit