webview css: Give code blocks the same background as in webapp.
The parent commit copied the CSS code we use for this in the webapp,
but there was one difference: it changed the background-color used in
dark mode, to avoid unsightly white corner tabs which mysteriously
don't appear in the webapp.

Turns out the webapp has the following block, in
rendered_markdown.scss :

  .codehilite {
      display: block !important;
      border: none !important;
      background: none !important;
  }

That causes the `.codehilite { background-color: hsl(0, 0%, 98%); }`
in turn to have no effect.

The `!important` seems like an unnecessary hack, but we can simply
delete that block that it disables.  Do that, and then restore the
specific background color for dark mode to match the webapp.