flags reducer: Combine `mentioned` and `wildcard_mentioned` flags.
As in the webapp, when we use the `mentioned` flag, we always mean
`mentioned || wildcard_mentioned` -- see `static/js/message_store.js
-> set_message_booleans` in `zulip/zulip`.

This commit combines the flags `mentioned` and `wildcard_mentioned`
such that if `wildcard_mentioned` flag is set, a corresponding entry
will be added to `mentioned` state as well.

Also added tests for the same.