notifications: Fix notifications reporting wrong number of conversations.
To create a notification, messages must be group by conversation.  The
sender id of a message was being used as one of the criteria for
these groups.  This was incorrect since a conversation can contain
messages from different senders.  The sender id criteria has been
removed from most of the grouping algorithm.  It is still used along with
the receiver's id in the case of private messages. Other methods have been
modified to work with the new groupings.  This commit partially addresses
issue #3591.  The information shown is now correct,  but hasn't yet been
reformatted as described in #3591.