android build: Bump compileSdkVersion to 29, aka Android 10 "Q".
This version is now stable and out of preview!  So it's time to
start using it in the build.  This will give us things like better
nullability annotations in the platform libraries, and the ability
to refer to Build.VERSION_CODES.Q as needed in conditionals.

This is not related to which versions we support for running the app.
See comments and the further info linked from there for details.