Android Notifications: Fix errors made in previous commit.
In the previous commit the class
object Pm
was changed to the data class
data class Pm (val senderIdString: String, val userIdString: String ).
This turns out to be unnecessary and also causes as syntax error in the
FcmMessageTest class.  This has been corrected.

Also the previous commit altered how stream messages were divided into
different conversations. They were further divided based on the topic. In
hind site, this may not be a good idea, since it will produce more
conversations, each of which will occupy an additional line in the
notification window.  Either way, it is out of scope for this or the
previous commit.  If it is decided to do this, it will be done in a
future commit.