SmartUrlInput: Refactor to consolidate logic.
Refactor SmartUrlInput.js and its associated utility functions.
No (intended) functional differences.

The key element of this refactor, and one of the original
motivating concerns, is that the logic for determining what
fragments to display is now almost entirely localized and unified
in `autocompleteRealmPieces`.

There is a small change in SmartUrlInput's own semantics: the
parameter formerly named `append` is now expected to be a proper
domain, with no prefixed period.  This has been propagated
appropriately.