docs/style: Begin a style-guide doc, with a point that came up today.
This is by no means the most important point of style; but I'd just
written about it in a PR comment (on #3533), and figured it should
have a more lasting home.  Lots more we can add later.

The list of examples from the Linux kernel of `Reported-by:` etc.
comes from the following command in a linux.git clone:

  $ git log --format=%b -n10000 --no-merges \
    | grep -Po '^[A-Za-z-]+: ' | sort | uniq -c | sort -n

That goes through a little over 2 months of history.