ComponentWithOverlay: Cut defaults where no sensible default exists.
The implicit BRAND_COLOR and 'top-right' just obscure what's going on
when looking at a call site where they aren't specified.

The default size of 0 isn't used anywhere, and wouldn't make a lot of
sense.  Perhaps a vestige of some previous version which used that as
a hack for what `showOverlay` now does?