narrowsReducer: Check `operation` match is exhaustive statically, too.
This would cause Flow to give an error if there were any way to reach
this case consistent with the stated types.

The line that throws an exception is still helpful too, because like
much of our data this action object comes largely from the server --
so there's no static guarantee that it actually *does* follow the
stated types.  (This is an area where we don't currently have much of
a "crunchy shell".)