docs: add instructions for installing the NDK
An updated version of the NDK is currently necessary for building
Zulip on a physical device. Note this, and provide instructions for
acquiring it.

(These are not as precise and detailed as they might be; they have
been written on the assumption that the user has either a) gone
through the "Getting Started" instructions recently, or b) has had
their setup in place long enough to update their emulator and SDK
through Android Studio.)