presence: Initialize lastReportPresence as "long ago", not "now".
Before we were setting it as `new Date()`. So when app was opened, this
variable was initialized as `new Date()` and we call `reportPresence`
from `fetchActions`. As diff is not >= 60, so `reportPresence` returns
without reporting to the server. And one cycle of reporting to server
was missed, which was resulting in delay marking user as active.

Now initialize it as `new Date(0)`, means last presence reported was
long time back, which will result in reporting promptly on app launch.

Fixes: #3590