webview build: add new script `tools/build-webview`
At present, when loading content into the WebView, we inject all
non-image assets as part of the generated HTML. This is more than a
bit wasteful, considering that large swaths of the supporting CSS and
JavaScript is entirely static from invocation to invocation;
additionally, it makes KaTeX support somewhere between "nontrivial"
and "impossible".

This commit creates a build subsystem which can inject static
webview-supporting assets into the application bundle, into a
`webview` directory.

We also create a stub directory for the subsystem to copy assets from;
at present this contains only a README.md file (which will *not* be
copied), but this will be expanded in the following series of commits.

Inspired by #2690. Assists with #2660, #3595, and potentially many
other issues.