android notif: Don't set initialNotification redundantly with event.
When we emit the notificationOpened event, it's because we've
determined the app is already running -- which means it won't see a
value we set on initialNotification.  So, do just one or the other of
those two things, whichever we determine is appropriate; not both.

In the source code, this looks like inlining emitOrLaunch into
notifyReact, and then sinking the `initialNotification` assignment
into two of the three exit paths but not the third.

The difference this causes in (internal) behavior is only in
the RESUMED case, where we were setting initialNotification and
no longer do.  This should have no user-observable effect.