webview css: inject interpolated constant directly into CSS
The current value of BRAND_COLOR is already being used at several
places within the stylesheet -- just with non-unitary opacities. If we
ever wanted to change this color, it would already be more work than
changing a single line in `constants.js`.

As such, we lose little by inlining it explicitly. This makes
`cssBase` entirely static.