reactions [nfc]: Simplify loop in aggregateReactions.
Two cheers for mutation!

Well, really: three cheers for the nice simple syntax traditionally
associated with mutation.  In this instance, `item` is always an
object we created within this very loop and which no other code can
have seen yet -- essentially, it's not even properly initialized until
this loop is done -- so it's safe to mutate it.  And writing the code
that way saves a good bit of goo and makes the logic cleaner to read.