xcode: Add a "scheme" for running a release-mode build.
This adds a new "scheme", which Apple's docs (in an old version --
somehow with Apple the highest-quality docs are always being abandoned
for hipper designs with less-informative text) define like so:

  An Xcode scheme defines a collection of targets to build, a
  configuration to use when building, and a collection of tests to
  execute.

  You can have as many schemes as you want, but only one can be active
  at a time. [...]  When you select an active scheme, you also select
  a run destination (that is, the architecture of the hardware for
  which the products are built).

  https://developer.apple.com/library/archive/featuredarticles/XcodeConcepts/Concept-Schemes.html

The scheme is the dropdown by the top left of the main Xcode window,
where we usually select "ZulipMobile".  (I think the "run destination"
is the dropdown just to its right, where we select a specific device
or emulator to run on, or else "Generic iOS Device" when building an
artifact to upload to the App Store.)

The new "ZulipMobile release-mode" is just like "ZulipMobile", except
that Cmd-r ("Run") will do a release build and run that on the target
device, rather than a debug build.

The "schemes" like the existing one and this new one live at a layer
of indirection above the "configurations", of which we have two, named
"Debug" and "Release".  They contain a very small amount of
configuration of their own, and mainly just a list of by-name
references to configurations.  The configurations themselves contain
quite a bit more detail.

Some helpful discussion:
  https://rhult.github.io/articles/testing-release-builds/

We're in "the simple case" in that post's terms: both our debug and
release configurations already use development certs for signing.
(For an actual release, I guess the artifact gets re-signed at upload
time?)

I originally did this in August, while debugging #3588; didn't commit
it, in part because I didn't feel I understood what it was doing.
This version is slightly cleaned up and I've added this commit
message, but it has the same substance as the config I left
uncommitted then.  Also I've read more and experimented more,
and I think I grasp at least some of what this means now. :)