metro [nfc]: Drop comment that made better sense in the template.
This was OK as long as the file was more or less verbatim from
upstream's template -- helps keep it aligned for the sake of future
upgrades.  But we're about to make substantial changes to this file,
which will both make the comment make less sense, and make that no
longer really helpful for upgrades.