rewriteImageUrls [nfc]: Use generic `toString()` over `.href`.
This is among the features that the MDN compatibility table says
arrived in Chrome 52, but that I discovered today (and sent upstream
a PR to the effect that) actually date at least to Chrome 37.  So
we can happily use it after all.

The name `href` here is a bit quirky and idiosyncratic.  It's not
obvious to me, without looking it up, that it should mean simply the
whole URL as a string, and not some particular twist on or massaging
of it.  Whereas `toString()` always means "convert this thing into a
string"; and for an object like a `URL` that is the parsed form of a
thing that canonically exists as a string, its only sensible meaning
is to serialize the thing in that canonical form.