rewriteImageUrls [nfc]: Say in jsdoc that the given element is included.
Simply "beneath" sounds like it might mean only the element's proper
descendants, excluding the element itself.  (In fact a comment inside
the implementation seems to use the same word with that meaning.)

While here, s/tag/element/ to conform to the jargon of the HTML spec,
and of the DOM API we're using.  The structured, already-parsed things
we're manipulating here are "elements"; a "tag" is a thing that exists
in the HTML syntax before it's parsed.  (For a "void element" like
`img` that can't have contents, the distinction doesn't hugely matter:
there was exactly one tag in the HTML per element in the DOM.  Still
good to be consistent, particularly with the API this code is
constantly using.)