auth [nfc]: Say "web auth" for what's only sometimes OAuth.
The thing the various cases of this codepath really have in common
isn't OAuth: even for the auth methods that happen to be OAuth
underneath, we don't touch any OAuth-related details about them.

(And indeed `remoteuser` auth is not OAuth or even much like it; and
SAML auth, which we've newly added on the server and will add here
soon, isn't OAuth either although it looks similar if you squint.)

Rather, what this codepath is really about is that the auth flow is
one we don't have a specific implementation of in the app.  To deal
with that, we hand the user a browser where they can go through it
on the web; there's a twist to have the server send them back to the
app when done.

So, give this code a more appropriately generic name.
Also write a comment explaining what it's about.