Emoji: Show emoji even if no longer active.
This component is told a specific emoji to display; it should
display it regardless of whether the emoji is still active and
available for new uses.

At present this is NFC because this component is only used in the
emoji picker screen and emoji autocomplete, and both of those use
only the active image emoji.

In the future we'll want to use this for showing emoji from past
reactions, which may no longer be active.  Then this difference
will become important.