sentry: update and fix embedded Xcode build scripts
Rewrite the Sentry-related build scripts.

Formerly we created and shipped Sentry artifacts on all builds not
targeting the iPhone Simulator. This is not really what we want (nor
previously wanted).

Instead, we now check if a) we're doing a "Release" build, and b) the
current commit has a Git tag. If both of these are true, we go through
sentry-cli to produce and upload build artifacts.