sentry: complete rewrite of all Sentry-invoking code (+ Flow types)
In the transition from 0.x to 1.0.x, The Sentry API has undergone a
complete overhaul. The old `raven-js` internals have gone away,
replaced by a unified JavaScript SDK that dispatches to native
backends where possible. Additionally Sentry has gone to some lengths
to unify their API across platforms.

This transition properly calls for a complete reanalysis and rewrite
of every part of our code that directly invokes Sentry. One should
perhaps think of this not as an upgrade so much as a movement to a
completely different SDK.

Unfortunately, this was not realized initially, during the (partial!)
upgrade performed in commit ec101ddcb. The new API is in many ways
only _subtly_ incompatible with the old -- and although TypeScript
type definitions are provided, Flow type definitions are not. The old
code in place appeared to work at first, but was ultimately broken.

(The exact symptoms of this are not known, and probably aren't
relevant; as the broken partial update hasn't made it out to a public
release, we won't need to worry about sifting through partial error
reports.)