webview [nfc]: Remove contentious comment; isolate its subject
Having this comment block in place has already cost us more time and
effort than it could ever reasonably save. It is a _fractal can of
worms._ Remove all attempts at explanation from it.

Since it's been firmly established to be useless for all other
purposes, move the value to its use-site, instead of placing it near
the top of the file where it's more likely to distract people.