flow-typed: force use of custom Sentry libdef
In commit 6219e5a23, it was verified that Flow used the definitions
present in the custom libdef for `react-native-document-picker`,
ignoring the versions in `flow-typed/npm` without further
configuration.

I have been able to reproduce this verification... and yet, it does
not hold true for the custom `@sentry/react-native` libdef. It is
ignored in favor of the one in `npm/@sentry/react-native`.