webview build: add no-sanity-checking option
Add option to disable sanity checks for webview directory.

These checks are mostly present to raise targeted alarms if something
goes wrong with the build system after a change or upgrade. There's no
need to have them in place if we're invoking the script manually for
testing purposes, though.

(No functional changes when this option is not used.)